Fixed documentation of activation energy in LL surface growth rate.
I believe the activation energy of 21100 in the growth rate is incorrect based on the cited paper (Leung et al. 1991, "A simplified reaction mechanism for soot formation in nonpremixed flames"). I believe the activation energy should be 12100. The code implementation in src/growthModels/growthModel_LL.cc has an activation energy of 12100 and looks correct to me, so this is merely a documentation typo.